Update AVA to 1.4.1 -> 2.4.0 -> 3.11.1
AVA 1.x and 2.x had `meow` as a dependency, which in turn depended on a
wildly outdated version of yargs-parser. 3.0 replaced `meow` with a
direct dependency on `yargs` 15.1.0 (yp-16.1.0), 3.6 with 15.3.1
(18.1.1) then 3.10 with 15.4.0 (18.1.2) which meets the patch.

As of 2.0, dirs named 'helpers' are no longer given special treatment,
so we need to manually define the path. Fixtures are also no longer
ignored, so to fix that we add an exception in the files config.

3.0 once again ignores fixtures and helpers dirs by default, but will
not compile helpers. For that we need @ava/babel to replace the Babel
functionality that was stripped from AVA's core.